https://issues.apache.org/activemq/browse/AMQCPP-272
Add new CMS exception to the Windows Project files and fix any new warnings found while compiling the code changes.

git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@890972 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/decaf/lang/Thread.cpp
@@ -613,7 +613,7 @@ Thread* Thread::currentThread() {
     // instance so that other threads in Decaf can join it and wait on it.
     if( result == NULL ) {
         result = Thread::createForeignThreadInstance(
-            std::string( "ForeignThread-" ) + Integer::toString( Thread::getId() ) );
+            std::string( "ForeignThread-" ) + Long::toString( Thread::getId() ) );
     }
 
     return (Thread*)result;
